requirements.txt
Django>=5.0,<6.0
firebase-admin>=6.5.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
psycopg[binary]
python-dateutil>=2.8
requests>=2.31
openai>=1.0